Add `child_spec/1` function to `TelemetryMetricsPrometheus.Core`
The `TelemetryMetricsPrometheus.Core` module has a `child_spec/1` function and a
`start_link/1` function. These functions start the registry, or add a
registry to the calling supervisor.

These functions can be used to add `TelemetryMetricsPrometheus.Core` to the
supervision tree of the calling application. The `init/2` function has been
deprecated.

The functions for registering metrics and monitoring the table have been moved
into the registry, however their functionality remains the same.

The metrics can be passed in as an option, this reflects how the
`Telemetry.Metrics.ConsoleReporter` module works.