Update child_spec/1 to manually construct child spec
The child_spec is now created manually so that the registry isn't exposed as the
id. The registry is still started as a result of using the child_spec.

The wait for the tests has been updated to use the `start_link/1` function and
`:sys.get_state/1` to ensure the registry is initialized.

Metrics are now a required parameter in both the core and the registry.