Move buckets entry into reporter_options
Previously, the `:buckets` was a key on the `Metrics.Distribution` struct, but has been removed with this commit: https://github.com/beam-telemetry/telemetry_metrics/commit/cc9b30952d36ac8506a2fba30b3905c990de0a7b. As the commit suggests moving this to the reporter_options, move the checks 'for the `:buckets` field to the check the `:reporter_options` keyword list.