Add reporter option to allow gauges for summed metrics (#36)
* Add reporter option to allow gauges for summed metrics

Some summed metrics can decrease as well as increase, in which case a
Prometheus counter isn't suitable as they are monotonic. A Prometheus
gauge fits nicely as it allows for arbitrary increases and decreases
value.

At the moment, it's only possible to get a gauge from `last_value`,
meaning if you want to report a sum goes up and down, you currently need
to sum the value yourself and make it available to `last_value` via
something like `telemetry_metrics_poller`.

This commit allows users to pass the `:prometheus_type` reporter option
when declaring a `sum`, which determines whether the metric exported to
prometheus is a counter or a gauge.